Remove schema4 and rename schema7 to schema
Removes the last references of schema4 and
renames schema7 to be the schema object.
